backend/pyproject.toml
[project]
name = "warehouse-management-backend"
version = "0.1.0"
description = "FastAPI backend for warehouse management system"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "supabase>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "qrcode[pil]>=7.4.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "httpx>=0.25.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]